core: move Comfort to PhysicsRollingStock.kt
