--- core/osrd-mp/src/commonMain/kotlin/fr/sncf/osrd/railjson/schema/rollingstock/Comfort.kt
@@ -1,7 +1,0 @@
-package fr.sncf.osrd.railjson.schema.rollingstock
-
-enum class Comfort {
-    AIR_CONDITIONING,
-    HEATING,
-    STANDARD,
-}
