+++ core/src/main/kotlin/fr/sncf/osrd/stdcm/graph/EnvelopeSimContextBuilder.kt
@@ -1,9 +1,9 @@
 package fr.sncf.osrd.stdcm.graph
 
+import fr.sncf.osrd.envelope_sim.Comfort
 import fr.sncf.osrd.envelope_sim.EnvelopeSimContext
 import fr.sncf.osrd.envelope_sim.PhysicsRollingStock
 import fr.sncf.osrd.path.interfaces.PhysicsPath
-import fr.sncf.osrd.railjson.schema.rollingstock.Comfort
 
 /** Computes the rolling stock effort curves that will be used and creates a context */
 fun build(
