+++ core/src/main/kotlin/fr/sncf/osrd/stdcm/graph/STDCMGraph.kt
@@ -4,11 +4,11 @@ import fr.sncf.osrd.api.FullInfra
 import fr.sncf.osrd.api.S3Context
 import fr.sncf.osrd.api.STDCMTimetableData
 import fr.sncf.osrd.envelope.Envelope
+import fr.sncf.osrd.envelope_sim.Comfort
 import fr.sncf.osrd.envelope_sim.allowances.AllowanceValue
 import fr.sncf.osrd.envelope_sim.allowances.AllowanceValue.FixedTime
 import fr.sncf.osrd.graph.Graph
 import fr.sncf.osrd.pathfinding.constraints.CachedBlockConstraintCombiner
-import fr.sncf.osrd.railjson.schema.rollingstock.Comfort
 import fr.sncf.osrd.sim_infra.api.ZoneId
 import fr.sncf.osrd.sim_infra.impl.TemporarySpeedLimitManager
 import fr.sncf.osrd.stdcm.CachedBlockMaxSpeedEnvBuilder
