+++ core/src/test/kotlin/fr/sncf/osrd/stdcm/BacktrackingTests.kt
@@ -1,7 +1,7 @@
 package fr.sncf.osrd.stdcm
 
 import com.google.common.collect.ImmutableMultimap
-import fr.sncf.osrd.railjson.schema.rollingstock.Comfort
+import fr.sncf.osrd.envelope_sim.Comfort
 import fr.sncf.osrd.sim_infra.api.Block
 import fr.sncf.osrd.stdcm.infra_exploration.BlockLocation
 import fr.sncf.osrd.stdcm.preprocessing.OccupancySegment
