+++ core/src/test/kotlin/fr/sncf/osrd/stdcm/STDCMHelpers.kt
@@ -4,7 +4,7 @@ import com.google.common.collect.ImmutableMultimap
 import fr.sncf.osrd.api.FullInfra
 import fr.sncf.osrd.conflicts.SpacingRequirement
 import fr.sncf.osrd.envelope.Envelope
-import fr.sncf.osrd.railjson.schema.rollingstock.Comfort
+import fr.sncf.osrd.envelope_sim.Comfort
 import fr.sncf.osrd.sim_infra.api.*
 import fr.sncf.osrd.sim_infra.impl.TemporarySpeedLimitManager
 import fr.sncf.osrd.standalone_sim.EnvelopeStopWrapper
